requirements.txt
streamlit
streamlit_option_menu
google.generativeai 